api/requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
redis
python-dotenv